+++ src/components/pages/contacts/contactsCards/index.ts
@@ -1,0 +1,1 @@
+export * from './ContactsCards';
