+++ src/components/pages/slugPageBreadcrumbs/SlugPageBreadcrumbs.tsx
@@ -1,0 +1,25 @@
+import Breadcrumbs from '@/components/global/Breadcrumbs/Breadcrumbs';
+import { useTranslations } from 'next-intl';
+import { FC } from 'react';
+
+type SlugPageBreadcrumbsProps = {
+  title: string;
+};
+
+export const SlugPageBreadcrumbs: FC<SlugPageBreadcrumbsProps> = ({
+  title,
+}) => {
+  const t = useTranslations('Breadcrumbs');
+  const breadcrumbsLinks = [
+    {
+      name: t('homePage'),
+      url: '/',
+    },
+    {
+      name: title,
+      url: '',
+    },
+  ];
+
+  return <Breadcrumbs links={breadcrumbsLinks} />;
+};
