+++ src/components/pages/slugPageBreadcrumbs/index.ts
@@ -1,0 +1,1 @@
+export * from './SlugPageBreadcrumbs';
