+++ src/pages/[slug].tsx
@@ -1,4 +1,4 @@
-import Breadcrumbs from '@/components/global/Breadcrumbs/Breadcrumbs';
+import { SlugPageBreadcrumbs } from '@/components/pages/slugPageBreadcrumbs';
 import { SectionRenderer } from '@/components/sections/SectionRenderer';
 import { customRestApi } from '@/services/wpCustomApi';
 import { StyledHeaderWrapper, Title } from '@/styles/components';
@@ -6,7 +6,6 @@ import { SectionsType } from '@/types/components/sections';
 import { PageDataFullType, PageDataItemType } from '@/types/services';
 import { validateWpPage } from '@/utils/zodValidators/validateWpPage';
 import { GetServerSideProps, GetServerSidePropsContext } from 'next';
-import { useTranslations } from "next-intl";
 
 export const getServerSideProps: GetServerSideProps = async (
   context: GetServerSidePropsContext
@@ -24,7 +23,7 @@ export const getServerSideProps: GetServerSideProps = async (
     }
 
     const isValidSectionsData = validateWpPage(responseData);
-    if (!isValidSectionsData) throw new Error("Invalid SectionsData data");
+    if (!isValidSectionsData) throw new Error('Invalid SectionsData data');
 
     if (responseData && responseData.data) {
       const pageResponseData = responseData.data as PageDataFullType;
@@ -44,10 +43,10 @@ export const getServerSideProps: GetServerSideProps = async (
 
     return { notFound: true };
   } catch (error) {
-    console.error("Server Error:", error);
+    console.error('Server Error:', error);
     return {
       redirect: {
-        destination: "/500",
+        destination: '/500',
         permanent: false,
       },
     };
@@ -60,23 +59,11 @@ interface PageProps {
 }
 
 const SlugPage = ({ page, sections }: PageProps) => {
-  const t = useTranslations("Breadcrumbs");
-  const breadcrumbsLinks = [
-    {
-      name: t("homePage"),
-      url: "/",
-    },
-    {
-      name: page?.title,
-      url: "",
-    },
-  ];
-
   return (
     <>
       <StyledHeaderWrapper>
-        <Breadcrumbs links={breadcrumbsLinks} />
-        <Title as={"h1"} uppercase>
+        <SlugPageBreadcrumbs title={page.title} />
+        <Title as={'h1'} uppercase>
           {page?.title}
         </Title>
       </StyledHeaderWrapper>
